+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+ - Added hf felica rdunencrypted (@7homasSutter)
+ - Added hf felica rqresponse (@7homasSutter)
+ - Added hf felica rqservice (@7homasSutter)
+ - Added polling for felica standard (@7homasSutter)
  - Added lf t55xx dump save and lf t55xx restore for .bin and .eml files (@mwalker33)
  - Added lf t55xx detected to try without password first (@mwalker33)
  - Chg `lf indala read` - added indala 26bit decoding (@martinbeier)